pyproject.toml
[tool.poetry]
name = "netcup-easter-notifier"
version = "0.1.0"
description = ""
authors = ["Thimo Koenig <thimo.koenig@gmail.ch>"]
readme = "README.md"
packages = [{include = "netcup_easter_notifier"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.2"
schedule = "^1.2.0"
beautifulsoup4 = "^4.12.2"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
